# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "LetScrum API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3999",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/v1/members/{libraryId}/{memberUserId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "libraryId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "memberUserId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "libraryId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "memberUserId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "isAdmin": {
                  "example": "any"
                },
                "nickname": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/members/{libraryId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "libraryId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/members/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "libraryId": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                },
                "nickname": {
                  "example": "any"
                },
                "isAdmin": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/members/{libraryId}/{memberUserEmail}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "libraryId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "memberUserEmail",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}